+++ phpmyfaq/admin/assets/src/api/index.js
@@ -2,6 +2,7 @@ export * from './category';
 export * from './faqs';
 export * from './glossary';
 export * from './group';
+export * from './statistics';
 export * from './tags';
 export * from './user';
 export * from './news';