+++ phpmyfaq/assets/src/utils/cookie-consent.js
@@ -13,82 +13,71 @@
  * @since     2023-01-21
  */
 
-import 'vanilla-cookieconsent';
-
-const cc = initCookieConsent();
+import * as cc from 'vanilla-cookieconsent';
 
 cc.run({
-  current_lang: 'en',
-  autoclear_cookies: true, // default: false
-  page_scripts: true, // default: false
-  cookie_name: 'phpmyfaq_cc_cookie',
-  auto_language: 'browser',
-  // mode: 'opt-in'                          // default: 'opt-in'; value: 'opt-in' or 'opt-out'
-  // delay: 0,                               // default: 0
-  // autorun: true,                          // default: true
-  // force_consent: false,                   // default: false
-  // hide_from_bots: true,                   // default: true
-  // remove_cookie_tables: false             // default: false
-  // cookie_expiration: 182,                 // default: 182 (days)
-  // cookie_necessary_only_expiration: 182   // default: disabled
-  // cookie_domain: location.hostname,       // default: current domain
-  // cookie_path: '/',                       // default: root
-  // cookie_same_site: 'Lax',                // default: 'Lax'
-  // use_rfc_cookie: false,                  // default: false
-  // revision: 0,                            // default: 0
+  // root: 'body',
+  autoShow: true,
+  // disablePageInteraction: true,
+  // hideFromBots: true,
+  mode: 'opt-in',
+  // revision: 0,
 
-  onFirstAction: (user_preferences, cookie) => {
-    // callback triggered only once
+  cookie: {
+    name: 'phpmyfaq_cc_cookie',
+    // domain: location.hostname,
+    // path: '/',
+    // sameSite: "Lax",
+    expiresAfterDays: 182,
+  },
+  guiOptions: {
+    consentModal: {
+      layout: 'box inline',
+      position: 'top center',
+      equalWeightButtons: true,
+      flipButtons: false,
+    },
+    preferencesModal: {
+      layout: 'box',
+      equalWeightButtons: true,
+      flipButtons: false,
+    },
   },
 
-  onAccept: (cookie) => {
-    // ...
+  onFirstConsent: ({ cookie }) => {},
+
+  onConsent: ({ cookie }) => {},
+
+  onChange: ({ changedCategories, changedServices }) => {},
+
+  onModalReady: ({ modalName }) => {},
+
+  onModalShow: ({ modalName }) => {},
+
+  onModalHide: ({ modalName }) => {},
+
+  categories: {
+    necessary: {
+      enabled: true, // this category is enabled by default
+      readOnly: true, // this category cannot be disabled
+    },
   },
 
-  onChange: (cookie, changed_preferences) => {
-    // ...
-  },
-
-  languages: {
-    en: {
-      consent_modal: {
-        title: 'phpMyFAQ use cookies!',
-        description:
-          'Hi, phpMyFAQ uses essential cookies to ensure its proper operation and tracking cookies to understand how you interact with it. The latter will be set only after consent. <button type="button" data-cc="c-settings" class="cc-link">Let me choose</button>',
-        primary_btn: {
-          text: 'Accept all',
-          role: 'accept_all',
-        },
-        secondary_btn: {
-          text: 'Reject all',
-          role: 'accept_necessary',
-        },
-      },
-      settings_modal: {
-        title: 'Cookie preferences',
-        save_settings_btn: 'Save settings',
-        accept_all_btn: 'Accept all',
-        reject_all_btn: 'Reject all',
-        close_btn_label: 'Close',
-        cookie_table_headers: [{ col1: 'Name' }, { col2: 'Domain' }, { col3: 'Expiration' }, { col4: 'Description' }],
-        blocks: [
-          {
-            title: 'Cookie usage 📢',
-            description:
-              'I use cookies to ensure the basic functionalities of the website and to enhance your online experience. You can choose for each category to opt-in/out whenever you want. For more details relative to cookies and other sensitive data, please read the full <a href="./privacy.html" class="cc-link">privacy policy</a>.',
-          },
-          {
-            title: 'Strictly necessary cookies',
-            description:
-              'These cookies are essential for the proper functioning of my website. Without these cookies, the website would not work properly',
-            toggle: {
-              value: 'necessary',
-              enabled: true,
-              readonly: true,
-            },
-          },
-        ],
-      },
+  language: {
+    default: 'en',
+    autoDetect: 'document',
+    translations: {
+      de: './translations/cookie-consent/de.json',
+      en: './translations/cookie-consent/en.json',
+      pl: './translations/cookie-consent/pl.json',
     },
   },
 });
+
+const cookiePreferences = document.getElementById('showCookieConsent');
+if (cookiePreferences) {
+  cookiePreferences.addEventListener('click', function (event) {
+    event.preventDefault();
+    cc.showPreferences();
+  });
+}